split : split structure cmd and services
