+++ cmd/main.go
@@ -1,33 +1,9 @@
 package main
 
-import (
-	"fmt"
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-func setupRouter() *gin.Engine {
-	router := gin.Default()
-	router.SetTrustedProxies([]string{"192.186.1.2"})
-
-	// Root
-	router.GET("/", func(c *gin.Context) {
-		c.JSON(http.StatusOK, gin.H{
-			"message": "Hello, world!",
-		})
-		fmt.Printf("ClientIP: %s\n", c.ClientIP())
-	})
-
-	// Ping Test
-	router.GET("/ping", func(c *gin.Context) {
-		c.String(200, "pong")
-	})
-
-	return router
-}
+import "go-calendar-practice/go-calendar-practice/go-calendar/services"
 
 func main() {
-	router := setupRouter()
-	router.Run(":3000")
+	const PORT = ":3000"
+
+	services.LaunchServer(PORT)
 }
